Move GitHub concerns into GitHub module
