+++ metashare/api/tests/jobs.py
@@ -3,13 +3,11 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 from django.utils.timezone import now
-from github3.exceptions import UnprocessableEntity
 from simple_salesforce.exceptions import SalesforceGeneralError
 
 from ..jobs import (
     _create_branches_on_github,
     _create_org_and_run_flow,
-    _try_to_make_branch,
     commit_changes_from_org,
     create_branches_on_github_then_create_scratch_org,
     delete_scratch_org,
@@ -78,46 +76,6 @@ class TestCreateBranchesOnGitHub:
             )
 
             assert not repository.create_branch_ref.called
-
-    def test_try_to_make_branch__duplicate_name(self, user_factory, task_factory):
-        repository = MagicMock()
-        resp = MagicMock(status_code=422)
-        resp.json.return_value = {"message": "Reference already exists"}
-        repository.create_branch_ref.side_effect = [UnprocessableEntity(resp), None]
-        branch = MagicMock()
-        branch.latest_sha.return_value = "1234abc"
-        repository.branch.return_value = branch
-        result = _try_to_make_branch(
-            repository, new_branch="new-branch", base_branch="base-branch"
-        )
-
-        assert result == "new-branch-1"
-
-    def test_try_to_make_branch__long_duplicate_name(self, user_factory, task_factory):
-        repository = MagicMock()
-        resp = MagicMock(status_code=422)
-        resp.json.return_value = {"message": "Reference already exists"}
-        repository.create_branch_ref.side_effect = [UnprocessableEntity(resp), None]
-        branch = MagicMock()
-        branch.latest_sha.return_value = "1234abc"
-        repository.branch.return_value = branch
-        result = _try_to_make_branch(
-            repository, new_branch="a" * 100, base_branch="base-branch"
-        )
-
-        assert result == "a" * 98 + "-1"
-
-    def test_try_to_make_branch__unknown_error(self, user_factory, task_factory):
-        repository = MagicMock()
-        resp = MagicMock(status_code=422, msg="Test message")
-        repository.create_branch_ref.side_effect = [UnprocessableEntity(resp), None]
-        branch = MagicMock()
-        branch.latest_sha.return_value = "1234abc"
-        repository.branch.return_value = branch
-        with pytest.raises(UnprocessableEntity):
-            _try_to_make_branch(
-                repository, new_branch="new-branch", base_branch="base-branch"
-            )
 
 
 def test_create_org_and_run_flow():
